+++ html-conduit/ChangeLog.md
@@ -1,3 +1,7 @@
+## 1.2.0
+
+* Drop system-filepath
+
 ## 1.1.1.2
 
 * Fix a bug with double-unescaping of entities